Move get column read implementation into separate compilation unit
